requirements.txt
re
pyyaml
argparse
robobrowser